Add files via upload
This project is a Python application that allows you to control media playback on your computer using hand gestures captured from a webcam. By utilizing computer vision techniques with the Mediapipe library, the application detects and tracks the movement of your hand. Based on the finger count and hand positions, it performs various media control actions.
Features

    Forward Skip: Extend any one finger to skip forward .
    Backward Skip: Extend any two fingers to skip backward.
    Volume Up: Extend any three fingers to increase the volume.
    Volume Down: Extend any four fingers to decrease the volume.
    Pause/Play: Extend all five fingers to pause or resume playback.

Requirements

    Python 3.x
    OpenCV
    Mediapipe
    PyAutoGUI

How to Use

    Install the required dependencies mentioned.
    Run the project_media_control.py script.
    Position your hand in front of the webcam.
    Perform the corresponding gesture to control media playback.

Please ensure that your webcam is properly configured and visible to the script.